Completed_Modified_Time_Series_Generator
Modified time series pipeline to export variant combinations in a string separated by commas instead of the textual representation of a python list. This allows the variant combinations to be searched and processed by other lists.

Minor edits were also made to the print statements in the time series pipeline to reduce the clutter in screen output.